+++ tencentcloud/resource_tc_as_scaling_config_test.go
@@ -103,9 +103,12 @@ func testAccCheckAsScalingConfigExists(n string) resource.TestCheckFunc {
 			return fmt.Errorf("auto scaling configuration id is not set")
 		}
 		asService := AsService{client: testAccProvider.Meta().(*TencentCloudClient).apiV3Conn}
-		_, _, err := asService.DescribeLaunchConfigurationById(ctx, rs.Primary.ID)
+		_, has, err := asService.DescribeLaunchConfigurationById(ctx, rs.Primary.ID)
 		if err != nil {
 			return err
+		}
+		if has < 1 {
+			return fmt.Errorf("auto scaling configuration not exists: %s", rs.Primary.ID)
 		}
 		return nil
 	}
@@ -123,8 +126,11 @@ func testAccCheckAsScalingConfigDestroy(s *terraform.State) error {
 			continue
 		}
 
-		_, _, err := asService.DescribeLaunchConfigurationById(ctx, rs.Primary.ID)
-		if err == nil {
+		_, has, err := asService.DescribeLaunchConfigurationById(ctx, rs.Primary.ID)
+		if err != nil {
+			return err
+		}
+		if has > 0 {
 			return fmt.Errorf("auto scaling configuration still exists: %s", rs.Primary.ID)
 		}
 	}