+++ tencentcloud/resource_tc_as_scaling_group_test.go
@@ -139,9 +139,12 @@ func testAccCheckAsScalingGroupExists(n string) resource.TestCheckFunc {
 			return fmt.Errorf("auto scaling group id is not set")
 		}
 		asService := AsService{client: testAccProvider.Meta().(*TencentCloudClient).apiV3Conn}
-		_, _, err := asService.DescribeAutoScalingGroupById(ctx, rs.Primary.ID)
+		_, has, err := asService.DescribeAutoScalingGroupById(ctx, rs.Primary.ID)
 		if err != nil {
 			return err
+		}
+		if has < 1 {
+			return fmt.Errorf("auto scaling group not exists: %s", rs.Primary.ID)
 		}
 		return nil
 	}
@@ -159,8 +162,11 @@ func testAccCheckAsScalingGroupDestroy(s *terraform.State) error {
 			continue
 		}
 
-		_, _, err := asService.DescribeAutoScalingGroupById(ctx, rs.Primary.ID)
-		if err == nil {
+		_, has, err := asService.DescribeAutoScalingGroupById(ctx, rs.Primary.ID)
+		if err != nil {
+			return err
+		}
+		if has > 0 {
 			return fmt.Errorf("auto scaling group still exists: %s", rs.Primary.ID)
 		}
 	}